backend/requirements.txt
numpy<2
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pymongo==4.13.2
opencv-python==4.11.0.86
ultralytics==8.3.159
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
requests==2.32.4
pillow==11.2.1
python-multipart==0.0.20
